requirements.txt
-e .
aws-cdk.core==1.102.0
aws-cdk.aws-stepfunctions==1.102.0
aws-cdk.aws-stepfunctions-tasks==1.102.0